Remove dapps fetcher => all in singleton dapps store + utils
